use oneBehindFetch in rules
