--- packages/validator-query/lib/rules/index.js
@@ -1,26 +1,0 @@
-/**
- * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-const VALIDATOR_RULES_URL = 'https://cdn.ampproject.org/v0/validator.json';
-
-module.exports = function(opt) {
-  opt = opt || {};
-  opt.url = opt.url || VALIDATOR_RULES_URL;
-  if (typeof process !== 'undefined') {
-    return require('./node')(opt);
-  }
-  return require('./browser')(opt);
-};
