+++ packages/validator-query/package.json
@@ -22,6 +22,9 @@
   "bugs": {
     "url": "https://github.com/ampproject/amp-toolbox/issues"
   },
+  "dependencies": {
+    "@ampproject/toolbox-core": "^1.0.0-beta.3"
+  },
   "homepage": "https://github.com/ampproject/amp-toolbox/tree/master/packages/validator-query",
   "gitHead": "cb496f081c871a25dc73113ae1e1fccfb59b2732"
 }
